requirements.txt
psycopg2-binary
requests
fastapi
uvicorn
pandas